+++ src/Propper.js
@@ -1,254 +1,127 @@
-import is from 'is';
+import Is from 'is';
+import inspector, { validator } from '@wonderlandlabs/inspector';
+import { compact, popObject } from './utils';
 
-import Validator from './Validator';
-
-const GENERIC_FAIL_MSG = '(#name#) passed bad value #value#';
-
-const defaultOnBadData = (name, value, error) => {
-  if (error) {
-    throw new Error(error);
-  }
-  throw new Error(GENERIC_FAIL_MSG.replace('#name#', name).replace('#value#', value));
-};
-
-const popObject = (obj, field, def) => {
-  if (field in obj) {
-    const out = obj[field];
-    delete obj[field];
-    return out;
-  }
-  return def;
-};
-
-const stringValidator = name => new Validator('string', `${name} must be a string`).setName('stringValidator');
-const requiredValidator = name => new Validator(a => !a, `${name} is required`).setName('requiredValidator');
-
-export default class Propper {
-  constructor(ClassDef, options = {}) {
-    this.classDef = ClassDef;
-    this.options = options;
-  }
-
-  get IDENTITY() {
-    return 'PROPPER';
-  }
-
-  /**
-   * this adds an 'isValid()' method and an 'propErrors()' method.
-   * isValid() returns true or false; propErrors() returns an array
-   * of prop/error objects, or null if there are no errors.
-   *
-   * NOTE: the only properties that will be tested are those defined AFTER
-   * addIsValid() is called. So you can't call it at the last minute
-   * and expect it to work!
-   *
-   * @param validMethodName
-   * @param propErrors
-   * @returns {*}
-   */
-  addIsValid(validMethodName = 'isValid', propErrors = 'propErrors', validatorRegistry = '__validators') {
-    Object.assign(this, {
-      validMethodName, propErrors, validatorRegistry,
-    });
-
-    if (!Reflect.has(this.classDef.prototype)) {
-      Object.defineProperty(this.classDef.prototype, validatorRegistry, {
-        configurable: false,
-        enumerable: false,
-        get() {
-          if (!this[`${validatorRegistry}_`]) {
-            this[`${validatorRegistry}_`] = new Map();
-          }
-          return this[`${validatorRegistry}_`];
-        },
-      });
-
-      Object.defineProperty(this.classDef.prototype, validMethodName, {
-        configurable: false,
-        enumerable: false,
-        get() {
-          if (this[propErrors]) {
-            return false;
-          }
-          return true;
-        },
-      });
-
-      Object.defineProperty(this.classDef.prototype, propErrors, {
-        configurable: false,
-        enumerable: false,
-        get() {
-          const errors = [];
-          Array.from(this[validatorRegistry].keys()).forEach((prop) => {
-            const value = this[prop];
-            const validator = this[validatorRegistry].get(prop);
-            const error = validator.try(value);
-            if (error) {
-              errors.push({ prop, error });
-            }
-          });
-          if (errors.length) {
-            return errors;
-          }
-          return null;
-        },
-      });
-    }
-    return this;
-  }
-
-  withValidator(...args) {
-    if (args[0] instanceof Validator) {
-      this._validator = args[0];
-    }
-    this._validator = new Validator(...args);
-  }
-
-  addString(name, overrides = {}) {
-    let regexValidator;
-    const regex = popObject(overrides, 'regex');
-    const min = popObject(overrides, 'min', null);
-    const max = popObject(overrides, 'max', null);
-    let minValidator;
-    let maxValidator;
-    const regexErrorMessage = popObject(overrides, 'regexErrorMessage', 'Bad string pattern');
-    if (regex) {
-      regexValidator = new Validator(value => !regex.test(value), regexErrorMessage);
-    }
-    const failsWhen = popObject(overrides, 'failsWhen');
-    if (!is.null(min)) {
-      minValidator = new Validator(
-        str => str.length < min,
-        `"#value#" too short; ${name} must be at least ${min} characters`,
-      );
-    }
-    if (!is.null(max)) {
-      maxValidator = new Validator(
-        str => str.length > max,
-        `#value# too long; ${name} cannot be longer than ${max} characters`,
-      );
-    }
-
-    if (!Reflect.has(overrides, 'defaultValue')) {
-      overrides.defaultValue = '';
-    }
-
-    overrides.failsWhen = Validator.compound(
-      stringValidator(name),
-      regexValidator,
-      minValidator,
-      maxValidator,
-      failsWhen,
+const DEFAULT_DEFAULTS = name => ({
+  onInvalid(value, error) {
+    console.log(
+      `error: setting ${name}`,
+      error, value,
     );
+  },
+});
 
-    return this.addProp(name, overrides);
+class Propper {
+  constructor(BaseClass, options = {}, defaults = DEFAULT_DEFAULTS) {
+    Object.assign(this, options);
+    this.defaults = {};
+    this.BaseClass = BaseClass;
   }
 
   addProp(name, options = {}) {
-    let localName = null;
+    const defaults = Is.function(this.defaults) ? this.defaults(name) : this.defaults || {};
+    const def = { ...defaults, ...options };
+    const tests = compact([popObject(def, 'type', null), ...(popObject(def, 'tests', []))]);
+    const required = popObject(def, 'required', null);
+    const onChange = popObject(def, 'onChange');
+    const enumerable = popObject(def, 'enumerable', false);
+    let onInvalid = popObject(def, 'onInvalid');
 
-    const definition = Object.assign({}, options);
-    // optionally the initial value is set with a function to ensure unique references
-    localName = popObject(definition, 'localName', `_${name}`);
-
-    let validator = this._validator || null;
-
-    const failsWhen = popObject(definition, 'failsWhen');
-    const type = popObject(definition, 'type');
-    const errorMessage = popObject(definition, 'errorMessage', GENERIC_FAIL_MSG).replace('#name#', name);
-    const onBadData = popObject(definition, 'onBadData', defaultOnBadData);
-    const defaultValue = popObject(definition, 'defaultValue', null);
-    const required = popObject(definition, 'required');
-    const onChange = popObject(definition, 'onChange');
-
-    let getDefault = () => defaultValue;
-    if (is.function(defaultValue)) {
-      getDefault = defaultValue;
+    if (onInvalid === 'throw') {
+      onInvalid = (value, error) => {
+        const err = new Error(error);
+        err.value = value;
+        throw err;
+      };
     }
 
-    if (failsWhen) {
-      if (Array.isArray(failsWhen) || failsWhen instanceof Validator) {
-        validator = Validator.compound(validator, failsWhen);
-      } else {
-        validator = Validator.compound(validator, new Validator(failsWhen, errorMessage));
+    const defaultValue = popObject(def, 'defaultValue', null);
+    let defaultFn = defaultValue;
+    if (!Is.function(defaultValue)) {
+      defaultFn = () => defaultValue;
+    }
+    const localName = popObject(def, 'localName', `_${name}`);
+    let validation = false;
+    if (tests.length || required) {
+      switch (required) {
+        case true:
+          validation = validator(tests, { required });
+          break;
+
+        case false:
+          validation = validator(tests, { required });
+          break;
+
+        default:
+          validation = false;
       }
     }
 
-    if (type) {
-      validator = Validator.compound(validator, new Validator(type, `${name} must be a ${type}`));
-    }
-
-    Object.assign(definition, {
+    const propDef = {
+      configurable: false,
+      enumerable,
       get() {
-        if (!Reflect.has(this, localName)) {
-          this[localName] = getDefault();
+        if (!(localName in this)) {
+          this[localName] = defaultFn();
         }
         return this[localName];
       },
-    });
+      set(value) {
+        if (localName in this) {
+          if (this[localName] === value) {
+            return;
+          }
+        }
+        if (localName in this) {
+          const lastVal = this[localName];
+          this[localName] = value;
+          if (onChange) {
+            onChange.bind(this)(value, lastVal);
+          }
+        } else {
+          this[localName] = value;
+          if (onChange) {
+            onChange.bind(this)(value);
+          }
+        }
+      },
+    };
 
-    if (required) {
-      const rv = requiredValidator(name);
-      try {
-        validator = Validator.compound(validator, rv);
-      } catch (err) {
-        console.log('bad validator attempt:', rv, validator, this);
-      }
-    }
-
-    if (validator) {
-      Object.assign(definition, {
+    if (validation) {
+      Object.assign(propDef, {
         set(value) {
-          const error = validator.try(value);
-          if (error) {
-            if (!(onBadData.bind(this))(name, value, error)) {
+          if (localName in this) {
+            if (this[localName] === value) {
               return;
             }
           }
-
-          if (this[localName] === value) {
+          const error = validation(value);
+          if (error) {
+            onInvalid(value, error);
             return;
           }
-          if (onChange) {
-            if (typeof onChange === 'string') {
-              this[onChange](value, this[localName], name);
-            } else {
-              onChange.call(this, value, this[localName], name);
-            }
-          }
+          const lastVal = this[localName];
           this[localName] = value;
-        },
-      });
-
-      if (this.validatorRegistry) {
-        this.classDef.prototype[this.validatorRegistry].set(name, validator);
-      }
-    } else {
-      //    console.log('no validator');
-      Object.assign(definition, {
-        set(value) {
-          if (this[localName] === value) {
-            return;
-          }
           if (onChange) {
-            if (typeof onChange === 'string') {
-              this[onChange](value, this[localName], name);
-            } else {
-              onChange.call(this, value, this[localName], name);
-            }
+            onChange.bind(this)(value, lastVal);
           }
-          this[localName] = value;
         },
       });
     }
 
-    /*
-    console.log('overrides:', overrides);
-    console.log('definition:', definition);
-    console.log('validator: ', validator);
-    */
 
-    Object.defineProperty(this.classDef.prototype, name, definition);
+    Object.defineProperty(this.BaseClass.prototype, name, propDef);
+
     return this;
   }
+
+  set defaults(value) {
+    this._defaults = { ...(value || {}) };
+  }
+
+  get defaults() {
+    return this._defaults;
+  }
 }
+
+export default Propper;