+++ src/index.js
@@ -1,10 +1,8 @@
 import Propper from './Propper';
-import easyPropper from './EasyPropper';
-import Validator from './Validator';
 
-const propper = (classDef, options = {}) => new Propper(classDef, options);
+const propInjector = (classDef, options = {}) => new Propper(classDef, options);
 
-export default propper;
+export default propInjector;
 
-export { Propper, propper, easyPropper, Validator };
+export { Propper, propInjector };