+++ CHANGELOG.md
@@ -11,3 +11,9 @@ adding metadata to package.json for NPM
 adding a type property for more semantic type checking
 executing onBadData in the context of the class
 allowing bad data to be set if onBadData returns true.
+
+## 2.0
+
+exporting validation functionality to a separate library, 
+`@wonderlandlabs/inspector`; importing into `@wonderlandlabs/propper`
+